backend/requirements.txt
django>=5.2,<6.0
djangorestframework
django-filter
django-cors-headers
drf-spectacular
django-allauth
dj-rest-auth[with_social]
djangorestframework-simplejwt
python-dotenv
PyJWT
cryptography
